sample_app/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
transformers>=4.37.0
torch>=2.2.0
Pillow>=10.0.0
bitsandbytes>=0.41.0
accelerate>=0.27.0 